+++ tests/unit/plugins/modules/test_na_ontap_security_ssh.py
@@ -1,4 +1,4 @@
-# (c) 2022, NetApp, Inc
+# (c) 2022-2025, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import (absolute_import, division, print_function)
@@ -6,17 +6,21 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 import pytest
+import sys
 
 import ansible_collections.netapp.ontap.plugins.module_utils.netapp as netapp_utils
 # pylint: disable=unused-import
-from ansible_collections.netapp.ontap.tests.unit.plugins.module_utils.ansible_mocks import patch_ansible, call_main
+from ansible_collections.netapp.ontap.tests.unit.plugins.module_utils.ansible_mocks import patch_ansible, call_main, create_module
 from ansible_collections.netapp.ontap.tests.unit.framework.mock_rest_and_zapi_requests import patch_request_and_invoke, register_responses
 from ansible_collections.netapp.ontap.tests.unit.framework.rest_factory import rest_responses
 
-from ansible_collections.netapp.ontap.plugins.modules.na_ontap_security_ssh import main as my_main      # module under test
+from ansible_collections.netapp.ontap.plugins.modules.na_ontap_security_ssh \
+    import NetAppOntapSecuritySSH as my_module, main as my_main      # module under test
 
-if not netapp_utils.has_netapp_lib():
-    pytestmark = pytest.mark.skip('skipping as missing required netapp_lib')
+if not netapp_utils.HAS_REQUESTS and sys.version_info < (2, 7):
+    pytestmark = pytest.mark.skip(
+        'Skipping Unit Tests on 2.6 as requests is not available')
+
 
 SRR = rest_responses({
     'ssh_security': (200, {
@@ -37,6 +41,11 @@ SRR = rest_responses({
                     "diffie_hellman_group_exchange_sha256",
                     "diffie_hellman_group14_sha1"
                 ],
+                "host_key_algorithms": [
+                    "ecdsa_sha2_nistp256",
+                    "ssh_rsa",
+                    "ssh_ed25519"
+                ]
             }],
         "num_records": 1
     }, None),
@@ -89,6 +98,25 @@ def test_modify_security_ssh_algorithms_rest():
         "ciphers": ["aes256_ctr", "aes192_ctr"],
         "mac_algorithms": ["hmac_sha1", "hmac_sha2_512_etm"],
         "key_exchange_algorithms": ["diffie_hellman_group_exchange_sha256"],
+    }
+    assert call_main(my_main, DEFAULT_ARGS, module_args)['changed']
+    module_args.pop('vserver')
+    assert call_main(my_main, DEFAULT_ARGS, module_args)['changed']
+
+
+def test_modify_host_key_algorithms_rest():
+    ''' test modify host_key_algorithms '''
+    register_responses([
+        ('GET', 'cluster', SRR['is_rest_9_16_1']),
+        ('GET', 'security/ssh/svms', SRR['ssh_security']),
+        ('PATCH', 'security/ssh/svms/02c9e252-41be-11e9-81d5-00a0986138f7', SRR['empty_good']),
+        ('GET', 'cluster', SRR['is_rest_9_16_1']),
+        ('GET', 'security/ssh', SRR['ssh_security']),
+        ('PATCH', 'security/ssh', SRR['empty_good']),
+    ])
+    module_args = {
+        "vserver": "AnsibleSVM",
+        "host_key_algorithms": ["ecdsa_sha2_nistp256", "ssh_rsa"]
     }
     assert call_main(my_main, DEFAULT_ARGS, module_args)['changed']
     module_args.pop('vserver')
@@ -162,3 +190,18 @@ def test_module_error_no_svm_uuid():
     }
     error = call_main(my_main, DEFAULT_ARGS, module_args, fail=True)['msg']
     assert 'Error: no uuid found for the SVM' in error
+
+
+def test_partially_supported_options_rest():
+    ''' Test REST version error for parameters '''
+    register_responses([
+        ('GET', 'cluster', SRR['is_rest_9_10_1']),
+    ])
+    args = {
+        'host_key_algorithms': [
+            'ecdsa_sha2_nistp256',
+            'ssh_rsa'
+        ]
+    }
+    error = create_module(my_module, DEFAULT_ARGS, args, fail=True)['msg']
+    assert 'Minimum version of ONTAP for host_key_algorithms is (9, 16, 1)' in error